SDC_fruitProject/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
numpy>=1.24
tensorflow>=2.15
